Make player 2's timer in synch with player 1's
