+++ ChromeExtension/Frontend/game-play-screen2.html
@@ -34,8 +34,8 @@
                 </div>
             </div>
      
-            <script type="module" src="assets/js/timer2.js"></script>
             <script type="module" src="assets/js/game_play2.js"></script>
+            <script type="module" src="assets/js/timer2.js"></script>
         </div>
     </div>
 </body>
